Fix typecasting issue for workflow types (#12217)
* Fix typecasting issue for workflow types

This addresses a number of the issues identified by CodeQL around
typecasting. We use 32 bit integers for these, so let's make sure we
parse and cast then as those everywhere it matters.


* Ensure to write to int into query


---------

+++ go/vt/vttablet/tabletmanager/vreplication/vreplicator.go
@@ -498,7 +498,7 @@ func (vr *vreplicator) setSQLMode(ctx context.Context, dbClient *vdbClient) (fun
 	if err != nil {
 		return resetFunc, err
 	}
-	if settings.WorkflowType == int32(binlogdatapb.VReplicationWorkflowType_OnlineDDL) {
+	if settings.WorkflowType == binlogdatapb.VReplicationWorkflowType_OnlineDDL {
 		vreplicationSQLMode = StrictSQLMode
 	}
